# swagger-collection.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Apartment Management API v1.0",
    "version": "1.0.0",
    "description": "OpenAPI 3 — import vào Swagger Editor / UI. Mọi JSON body dùng schema có `example` trên từng field. Các UUID trong ví dụ có dạng ngẫu nhiên (giống `Guid.NewGuid()` khi seed) và **không** phải id thật trong DB — sau khi seed, thay bằng giá trị từ API. JWT: login rồi gửi Authorization: Bearer. Seed data range: apartments BLK-0001..BLK-0150, bulk users demoblk00..demoblk39@seed.local, canonical accounts admin1/admin2 and user1..user6 @ apartment.local. UUID examples are placeholders; after seeding, lấy ID thật từ API list/detail."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/v1.0/Apartments/me": {
      "get": {
        "tags": [
          "Apartments"
        ],
        "summary": "My apartment",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Apartments/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "8f3e2c1a-9b5d-4f7e-a162-3c4d5e6f7089"
        }
      ],
      "get": {
        "tags": [
          "Apartments"
        ],
        "summary": "Get apartment by id (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Apartments"
        ],
        "summary": "Update apartment (Admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApartmentUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Apartments"
        ],
        "summary": "Soft delete apartment (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Apartments/{id}/restore": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "8f3e2c1a-9b5d-4f7e-a162-3c4d5e6f7089"
        }
      ],
      "post": {
        "tags": [
          "Apartments"
        ],
        "summary": "Restore apartment (Admin)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1.0/Apartments/{id}/images": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          },
          "example": "8f3e2c1a-9b5d-4f7e-a162-3c4d5e6f7089"
        }
      ],
      "post": {
        "tags": [
          "Apartments"
        ],
        "summary": "Upload apartment image (Admin, multipart)",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "JPEG/PNG/WebP image"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ApartmentUpdateDto": {
        "type": "object",
        "required": [
          "apartmentNumber",
          "floor",
          "area",
          "status"
        ],
        "properties": {
          "apartmentNumber": {
            "type": "string",
            "example": "BLK-0042"
          },
          "floor": {
            "type": "integer",
            "example": 7
          },
          "area": {
            "type": "number",
            "format": "double",
            "example": 58.54
          },
          "status": {
            "type": "integer",
            "description": "0=Available, 1=Occupied, 2=Maintenance",
            "example": 0
          },
          "description": {
            "type": "string",
            "nullable": true,
            "example": "Updated demo description 150"
          }
        }
      },
...
    }
  }
}